# Optimize histogram copying to lazily copy buckets