--- a/libnaucrates/include/naucrates/statistics/CLeftAntiSemiJoinStatsProcessor.h
+++ b/libnaucrates/include/naucrates/statistics/CLeftAntiSemiJoinStatsProcessor.h
@@ -24,7 +24,6 @@ namespace gpnaucrates
 			static
 			void JoinHistogramsLASJ
 					(
-					CMemoryPool *mp,
 					const CHistogram *histogram1,
 					const CHistogram *histogram2,
 					CStatsPredJoin *join_stats,