… CMakeLists.txt
… libnaucrates/src/statistics/CJoinStatsProcessor.cpp
… libnaucrates/src/statistics/CLeftAntiSemiJoinStatsProcessor.cpp
… libnaucrates/src/statistics/CLeftOuterJoinStatsProcessor.cpp
… libnaucrates/src/statistics/CProjectStatsProcessor.cpp
… libnaucrates/src/statistics/CStatistics.cpp
… libnaucrates/src/statistics/CStatisticsUtils.cpp
… libnaucrates/src/statistics/CUnionAllStatsProcessor.cpp
… server/src/unittest/dxl/statistics/CCardinalityTestUtils.cpp
… server/src/unittest/dxl/statistics/CHistogramTest.cpp
… server/src/unittest/dxl/statistics/CJoinCardinalityTest.cpp
… server/src/unittest/dxl/statistics/CMCVTest.cpp
… server/src/unittest/dxl/statistics/CStatisticsTest.cpp