Previously, when constructing new histograms, we would copy all buckets
from an existing histogram. This was expensive, both in optimization
time and memory, as this is used in PstatsDerive().

Now, we lazily copy the underlying bucket array, `m_histogram_buckets` only when modifying the individual buckets. We only modify the buckets on 3 occasions:

1. CapNDVs(): We set the distinct values of the bucket in some
cases.

2. NormalizeHistogram(): We set the frequency of the bucket to
another value.

3. DistributeBucketProperties(): We set both the frequency and distinct
values of the bucket.

In the vast majority of cases, we're simply copying this array and
reading from it; modifications are relatively rare, which makes this
much more efficient.

From profiling on my laptop, I saw TPC-DS Q72 memory utilization go from
700MB to 270MB and optimization time go from 5.5s to 4s.

The entire TPC-DS optimization time went from 67s to 59s (12%
improvement).

I'm open to suggestions on making this less error-prone. I manually went through places we referenced `m_histogram_buckets` and made sure we aren't modifying it and added a `const` modifier to some of the function CBucketArray parameters.

Authored-by: Chris Hajas <chajas@pivotal.io>