@chrishajas 2019-11-23
@hardikar Thanks for the review! I addressed 1 and 2, and created a new constructor without passing in a CBucketArray (and changed the callers where possible). Please take a look at the last 3 commits.