@chrishajas 2019-11-26
> Looks much better! Thanks!
> Leaving a note here for the 2 times:
> 
> 1. Use DeepCopyHistogramBuckets() in DistributeBucketProperties
> 2. There are many functions that use mp, instead of m_mp because it's passed in. We don't need to. Now's a good time to clean that up, but if in a rush, the current state is also good.

@hardikar Thanks! I've fixed both of those.